+++ NEWS.md
@@ -4,6 +4,7 @@ basis argument isn't specified.
 * Added calc_peak_info function for simple singlet analyses.
 * crop_spec function now maintains the orignal frequency scale.
 * The basis set used for analyses has now been added to the fit result object.
+* Bug fix for simulating basis sets with one element.
 
 # spant 0.4.0
 * Bug fix for SPAR/SDAT SVS voxel dimensions.